fixed bug number_of_atoms, preliminary version of static load imbalance test

+++ include/frame.hpp
@@ -73,7 +73,7 @@ class Frame {
     }
 	
 public:
-	long number_of_atoms; //store this for performance
+	size_t number_of_atoms; //store this for performance
 	
 	double t; // time information
